Extend slicer version handling to allow differentiation of Release vs Nightly
Slicer version should now be specified as a string of the form:

  Slicer X.Y[-Nightly]

Introduced the following functions:
* getSlicerMajorMinorVersion
* isSlicerReleaseVersion

Function assuming wiki page namespace was 'Nighty' have also updated to
extract the release identifier given the Slicer version. It means, the
pages in either one of these namespaces (Nightly, 4.4, ... ) can now
be updated based on the provided or auto-discoverted Slicer version.

Note that this is a backward incompatible change.