ENH: Addition of an option to choose the extension(s) to get stats of
The name of the extension was previously hardcoded in the script (SPHARM-PDM). Now the command line is parsed and the user can give one or multiple extension names and get the statistics for each of them.
An additional option can be given to print or not information while the script is running (--verbose).